+++ lib/src/repository/app_repository.dart
@@ -100,6 +100,12 @@ abstract class AppRepository {
     int limit = 20,
   });
 
+  Future<FeedEntryCheersUpdate> setFeedEntryCheers({
+    required String userId,
+    required String entryId,
+    required bool shouldCheer,
+  });
+
   Future<DrinkEntry> addDrinkEntry({
     required AppUser user,
     required DrinkDefinition drink,